+++ kernel/arch/x86/heap.h
@@ -5,7 +5,7 @@
 
 extern uintptr_t kheap_placement_address;
 extern uintptr_t kheap_init_address;
-extern uint kheap_end_address;
+extern uintptr_t kheap_end_address;
 
 uintptr_t kpmalloc_a(size_t);
 uintptr_t kpmalloc_p(size_t, uintptr_t*);
